Post and Get requests, using RxJava.
Added two central options- you can get a url, using HttpURLConnection,
by hitting the "Get" button. This fetches a simple HTML page as a
String, unformatted (the page is httpbin.org). Typing in the EditText
field causes POST requests to be sent (to httpbin.org/post), which in
turn fetches a page with the body of the POST request.

Bug: Typing too fast yields an error that crashes the app. In the
bicycling app, POST requests will all be either incremental (every ten
seconds or so) or scheduled (daily-ish), so I don't foresee this being a
large problem for the future. However, it is worth looking up on.